// src/app/client/admin/ad-body/userconfig/empinterface.ts
export class Empinterface {
    id: number = 0;
    profilepic: string = '';
    firstname: string = '';
    lastname: string = '';
    Ecode: number = 0;
    usertype: string = '';
    email: string = '';
    password: string = '';
    department: string = '';
    manager: string = '';
    
    
    
  
}